python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sentence-transformers==3.0.1
numpy>=1.26.0
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
groq>=0.9.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
gTTS>=2.5.0
pydantic>=2.0.0
httpx>=0.27.0
